Made a system for calculating exactly how much stamina will get used for the next run as you are running back to the bank.
It is linked to the Low energy threshold config setting - it will not allow your stamina to be <= the threshold.

Things it does:

* Takes into account the number of ticks it takes to get to the Blast Furnace (BF) and then back to the bank.

* Dynamically adjusts for weight based on what the player has equipped and what will be in the player's inventory during the next run.

* Calculates run energy required for each part of the run to and from the BF, including the 9 ticks to take ore to the BF, the 5 ticks you'll have running with an empty inventory (9 if it's a coal run), and finally the 4 ticks at the end when you will be holding bars.

The end result is pretty awesome. Even with the threshold set to 0, you won't run out of run energy, and stamina potion timers are used up as much as possible before another is highlighted.