Replaced Deposit Inventory message with relevant message for depositing ores or potions.
Added support for the Max Cape as it also provides the goldsmith gauntlets perk.

Fix for coal bag sometimes not showing as empty when it is.

Coal bag step accounts for 36 coal in bag now for GoldHybridMethod if wearing a smithing/max cape.

Added extra step to GoldHybridMethod and GoldMethod to make equipping the goldsmith gauntlets smoother.

Moved equipIceOrSmithsGloves to before waitForBars to encourage equipping them while running to the bars collection point.

Prevented null equipment variable exception.

Updated README to reflect changes.