Added checks to all methods for whether we need to leave the bank and collect bars.
Made oreOnConveyor actually mean the ore on the conveyor belt, and not ore on the conveyor or in the bar dispenser.

Added checks for whether ores are in the furnace alongside the check for ore on conveyor where appropriate.

Check if furnace has a metal bar before making the player wait for gold bars. If there's no metal bar then they don't ever need to wait for the gold.

Moved bank step for gold to the top