Added recentlyEmptiedCoalBag to account for the tick where we have emptied the coal bag but there is not yet coal in the inventory.
Renamed and changed description for tick perfect methods config item.

Re-added coal bag hack for when it sometimes shows that it's full/still has some coal when it does not.

Added extra condition for going back for gold bars when using a non-tick perfect method and in the bank.

Wait for gold bars when furnace has gold ore, not just when there is ore on the conveyor.

Got all methods working well.