Release/1.4.5 Potion Storage support, run energy calc update, max cape fix, coal bag & goldsmith gauntlets made optional. (#39)
Potion Storage is now supported. The plugin now knows when you have stamina/energy potions in the bank or in Potion Storage, and will highlight the potion if the potion is visible in the bank tab you have open. It works best when paired with bank tabs or inventory setups.

I updated the run energy calculation so the times the plugin suggests drinking a stamina dose will be more accurate.

I added all current variants of the max cape so now the plugin should get you to wear it as intended.

The coal bag and goldsmith gauntlets are now optional. The plugin will skip withdrawing/equipping them if you don't have them.

Fixed plugin thinking you should go to the bank instead of waiting for bars to be made.

Added Strange fruit to the options for energy regen.

Updated A Lot of deprecated imports. Took a while.